Convert testMakeHoleyWithSlash to unordered test fixes #5
JDK8, somehow, parses the fetch.txt file in a different order to
previous JDKs, resulting in a test for the first line in fetch.txt
failing. The BagIt specification, version 0.97, doesn't seem to say
anything about the order of the lines being important, so ignoring the
failure and instead converting the test to test all of the lines,
including a failure if any unrecognised lines are found.